refactor the code
