+++ bom/pom.xml
@@ -28,7 +28,7 @@
         <jackson.version>2.7.9</jackson.version>
         <msgpack.version>0.6.11</msgpack.version>
         <!--common-->
-        <httpcomponents.version>4.3.3</httpcomponents.version>
+        <httpcomponents.version>4.4.9</httpcomponents.version>
         <commons.fileupload.version>1.3.3</commons.fileupload.version>
         <!-- Log libs -->
         <slf4j.version>1.7.21</slf4j.version>
@@ -112,6 +112,12 @@
                 <groupId>org.apache.httpcomponents</groupId>
                 <artifactId>httpmime</artifactId>
                 <version>${httpcomponents.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.apache.httpcomponents</groupId>
+                <artifactId>httpclient</artifactId>
+                <version>4.5.5</version>
             </dependency>
 
             <dependency>
@@ -294,6 +300,37 @@
                 <groupId>org.apache.cxf</groupId>
                 <artifactId>cxf-core</artifactId>
                 <version>${cxf.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-lang3</artifactId>
+                <version>3.0</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.guava</groupId>
+                <artifactId>guava</artifactId>
+                <version>18.0</version>
+            </dependency>
+            <dependency>
+                <groupId>com.ecwid.consul</groupId>
+                <artifactId>consul-api</artifactId>
+                <version>1.3.1</version>
+            </dependency>
+            <dependency>
+                <groupId>javax.validation</groupId>
+                <artifactId>validation-api</artifactId>
+                <version>1.1.0.Final</version>
+            </dependency>
+            <dependency>
+                <groupId>joda-time</groupId>
+                <artifactId>joda-time</artifactId>
+                <version>2.9.9</version>
+            </dependency>
+            <dependency>
+                <groupId>com.pszymczyk.consul</groupId>
+                <artifactId>embedded-consul</artifactId>
+                <version>1.1.0</version>
+                <scope>test</scope>
             </dependency>
 
             <!-- alipay -->
