+++ core/common/src/main/java/com/alipay/sofa/rpc/common/utils/CommonUtils.java
@@ -16,10 +16,7 @@
  */
 package com.alipay.sofa.rpc.common.utils;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentMap;
 
 /**
@@ -275,5 +272,44 @@ public class CommonUtils {
             }
         }
         return sb.length() > 0 ? sb.substring(0, sb.length() - separator.length()) : StringUtils.EMPTY;
+    }
+
+    /**
+     * 将String数组转化为map
+     *
+     * @param pairs  目标数组
+     * @return HashMap
+     */
+    public static Map<String, String> toStringMap(String... pairs) {
+        Map<String, String> parameters = new HashMap<String, String>();
+        if (pairs.length > 0) {
+            if (pairs.length % 2 != 0) {
+                throw new IllegalArgumentException("pairs must be even.");
+            }
+            for (int i = 0; i < pairs.length; i = i + 2) {
+                parameters.put(pairs[i], pairs[i + 1]);
+            }
+        }
+        return parameters;
+    }
+
+    /**
+     * 比较集合是否一致，忽略顺序
+     *
+     * @param newCollection  新集合
+     * @param oldCollection 旧集合
+     * @return 是否一致
+     */
+    public static boolean isSameCollection(Collection<?> newCollection, Collection<?> oldCollection) {
+        if (newCollection == null || newCollection.isEmpty()) {
+            return true;
+        } else if (oldCollection != null) {
+            if (newCollection.containsAll(oldCollection) && oldCollection.containsAll(newCollection)) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+        return false;
     }
 }
