--- extension-impl/registry-consul/src/test/resources/ssl/ca.cert
@@ -1,26 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFNzCCAx+gAwIBAgIEM1EnWzANBgkqhkiG9w0BAQ0FADBFMQswCQYDVQQGEwJVSzEPMA0GA1UE
-BxMGTG9uZG9uMRIwEAYDVQQKEwlUZXN0T3JnQ0ExETAPBgNVBAMTCHRlc3Qub3JnMB4XDTE4MDcy
-NTA2MDkyMVoXDTI4MDcyMjA2MDkyMVowRTELMAkGA1UEBhMCVUsxDzANBgNVBAcTBkxvbmRvbjES
-MBAGA1UEChMJVGVzdE9yZ0NBMREwDwYDVQQDEwh0ZXN0Lm9yZzCCAiIwDQYJKoZIhvcNAQEBBQAD
-ggIPADCCAgoCggIBAMlb3p6euGnp+SWOnIVv0JklDeB6QyuUitgdx8dqgwJO1aJZG5feLO4alrKV
-gOZ+BLhD9Z8zQmFztEP36Bu16Acu2dLho1MTF4Nyul9da0cLMNm+WnorwB/Ef68XluIf9gEJgBS8
-nNPyLcWR6COQ23OPdE/jo9agAzb7pX4X9atNEjHmdwD4KKEVJaKqyE0pynNneDfJTS15GxVcGHoi
-qBcpwTSXmEzoCVein7YvJ9fCdXwteNO1fVvmKDdxtyIe8jgMYI7yVLRWfPfwMwf3NvCgB5sogYTw
-OIb2MZFFwTKK663s62U2xCtqNepI/GdtrmW7ThvnbYKki8VOyK9dmuOgsVxeWrEwGbzb+2kPlp8o
-xk03VcFxDE8XTRK/EO6qF+VYnmEQkbUVgtq6fMcWYGZeqRdcMnLJxfWCTjhmni3bTSsgpxLbu0z5
-4SNhFYIi/hAHyuatrGjRhiWmjmU5eDoem68Bdbxh6MgHFDdGzwgpQG+M2KimKXYYzicUdMm/WLx5
-hw6UTNvYyC8cMwZdhIYy7R96CnCqxQ6tW5I+rokTZy8F71bfdSE03rn2TFq6KpDRzBKYxa8BCjId
-W/nvRHvEwIJo5X8uwUhVObcebPEjVeX9HPr135INHBxGddu0qzao3mdbWrQYHDXX5IhN1MqEQAuX
-EuXDU0deX+CjezDPAgMBAAGjLzAtMAwGA1UdEwQFMAMBAf8wHQYDVR0OBBYEFEg8Tt7TRESjDuwj
-F6doAmhYu3H7MA0GCSqGSIb3DQEBDQUAA4ICAQBV2XmLkHIkSNT6KQRa1ehbH7I1FKOP199gQLqY
-Uu0ojQIlPwWrc1huOpoajKF5Wc/5EfAkTIFUFOF569KoRTyVrlINcSRCNkhbYdIHT8b+MZO/LAFX
-ICnbbDnqJG/Yqjqx6dz4Sd431GJvpYtA6kdygTRuj0j6QD2WgClXXl0l+Ue4TzJwqvGFm8oyjCB7
-b/4FS/qCbDHs+vdME8VFYIzPmaUpgRcg2f1T2NwADc2gCQ1cyWjWanskt76eSct2zVgrNFLbqtVd
-xchJTC5gFZBAqadgnutSyvU1BatkzauW4Vo8/ZMkz1QlMgTwfCdL+msPm/4XpLPpUbSe9GLED6BT
-9MCG5RsTDSwkEotaqjB+vcMAuivX6cmnNCY5ReX36iP31xgd5oiwqnPfhiGUFKJBKv8PiCl/ALzV
-IdaO/lO5thUvAlr8p6K/x2g918IDBb4LP9EjMFNwyJFBMSDkUrWk+o9VO761qYcshswkbg1TyAls
-FZdc4ILxM99nhIvD+D7q6iJjHPz9GxLmLpXvNE30JbFLsrUuccjBaUdE6b5xm8be83cAYrvRicgZ
-mPcr1hm5mKFYhUEOaa2zLWVvBMEQ4hnNLfSrH+2AVQFB/ryv7KDoeqU3R888g2ibQJvSXeUdXXq9
-uUloa5/qa0Lqysk8a2z369vM1VR0DbZ6Aus6vg==
------END CERTIFICATE-----
