--- extension-impl/registry-consul/src/test/resources/ssl/key.crt
@@ -1,71 +1,0 @@
-Bag Attributes
-    friendlyName: key
-    localKeyID: 54 69 6D 65 20 31 35 33 32 34 39 38 39 36 36 33 32 39 
-subject=/C=UK/ST=Unknown/L=London/O=TestOrg/OU=OU/CN=localBox
-issuer=/C=UK/L=London/O=TestOrgCA/CN=test.org
------BEGIN CERTIFICATE-----
-MIIFlTCCA32gAwIBAgIEWbbWyTANBgkqhkiG9w0BAQ0FADBFMQswCQYDVQQGEwJV
-SzEPMA0GA1UEBxMGTG9uZG9uMRIwEAYDVQQKEwlUZXN0T3JnQ0ExETAPBgNVBAMT
-CHRlc3Qub3JnMB4XDTE4MDcyNTA2MDkyNFoXDTI4MDcyMjA2MDkyNFowYjELMAkG
-A1UEBhMCVUsxEDAOBgNVBAgTB1Vua25vd24xDzANBgNVBAcTBkxvbmRvbjEQMA4G
-A1UEChMHVGVzdE9yZzELMAkGA1UECxMCT1UxETAPBgNVBAMTCGxvY2FsQm94MIIC
-IjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEAtpocEgeT814pH8C+7gDup5pf
-BkxBhOyxKvex/j/1Al4/D6lwThJlJ6NxGUKnKgt5jbr04O82NaTArYoTzOd2Yv+F
-PaF7nb/I7eRZziZkbNW2C3JBRktAxzpVxmVKrqH1n3zPZwS7Mc4LHSCjUwkr8g4G
-aMlW2Xi5mCapv5PCv2Jnu8D9xgwXrr4C4FdMn8++daJ4Ujk9qdhgWePXHb2/GvQN
-mEYLiopARpWmnfVGTOakkXL5ZmWdl9nESt9lowY3Nakoh+ntezc6E7FILqf8Frg7
-dhbkX5kuj/P0pCtbohkk1P4HHxZeFwlbxz2CCQRjMPxtjbIj/3AX6vZPMU7fsHLl
-3pqk9RkPxbJ41dX5vxV6mmV4AjSS5vGRRFwWmI8qB1GiCwrOZakp0iHZJWW7+4eE
-NCVac6GqQ85FlVifkvhhFcKqvIsD5Umh36H7gIQKiiukm5UEQVqlzVD+DeFiGpCy
-oOxEwqyJAsSS/d3sbnEcEuGtfnbEJytpBPDgOTJPS8ulOoASFBoPBsSQmpL9FIZV
-ysKhAlCeQ9KCL2v1gAfbzGf8fg2vclia3LaMo71rnmkc46kKBtzgdV+Nzp4XTtNK
-cJyt5jbVY6d/L0cX1bAI7MZR5r9DkXEp7jNlJInO+KP34DzgCpc+ts12rjOxWSpB
-gQe1CVNhnTileCqeO4sCAwEAAaNwMG4wHwYDVR0jBBgwFoAUSDxO3tNERKMO7CMX
-p2gCaFi7cfswLAYDVR0RBCUwI4IJbG9jYWxob3N0hwR/AAABhxAAAAAAAAAAAAAA
-AAAAAAABMB0GA1UdDgQWBBTlKZVdsFe6b0H80g2wSaZjZKYI/TANBgkqhkiG9w0B
-AQ0FAAOCAgEAfUTMeCqNjwprNfH+KbiQBZyTw/tjmRaPkDpQWOOc4zCNg0BthTkS
-f0DAj9urQmVV/Ki5lEOlpa2UwshG5/WHI4xIdODvilU3JtdYO9argn9I9r+VAVrr
-CAIlrt8iNXrQq2+tMwzhoKAS1GSE6fVZM8I2VzuiEqV97BDtJuIzqvUGK3Hy021H
-jU/3Us6r0eUwi0m+wnyuwaz/QYR0zoUfCT0JFvcynbIzd8uhGnnRoGACTqUh+6uc
-gtWQ1T5kZg9lDH7R/5eflS/UG93autSYeNWh9hpZfEvWMBuRhXLXjOV/M066Bvre
-lN5TKWDss098fzIsS4Dt2jsjosW537iR9fCpVPRz0b00TXS5xtzMWM5Bwo3Fqd4d
-YWHZKx3FVzBxt1ey/POrcTiw8jI6KdfhFSRSqJGvI9rOjM07qX/MYvrtrW1mgKBg
-4GGme9TE1PyylGA/3jN/z69aM2VcDd88W+Y754sPjy5rnMVbTy4q72pIxCWmc6ay
-vcacdQEf2IuEU8rjpkd/66MTeBMHhTHBE85AWH3dQArRVxGcc7mNvd3Ag3xS/dgy
-fG5/UZiYSeZEGCHuc6W9TJP7wzph91Uml9vCd/i+2I9Cguyte8dDhHmkITdsiYg2
-L/QiNzYdLDshBtazgVfCM7Z7o/WKB9rvqAW9Fdxmgi9gGgJqSrYvx3w=
------END CERTIFICATE-----
-Bag Attributes
-    friendlyName: CN=test.org,O=TestOrgCA,L=London,C=UK
-subject=/C=UK/L=London/O=TestOrgCA/CN=test.org
-issuer=/C=UK/L=London/O=TestOrgCA/CN=test.org
------BEGIN CERTIFICATE-----
-MIIFNzCCAx+gAwIBAgIEM1EnWzANBgkqhkiG9w0BAQ0FADBFMQswCQYDVQQGEwJV
-SzEPMA0GA1UEBxMGTG9uZG9uMRIwEAYDVQQKEwlUZXN0T3JnQ0ExETAPBgNVBAMT
-CHRlc3Qub3JnMB4XDTE4MDcyNTA2MDkyMVoXDTI4MDcyMjA2MDkyMVowRTELMAkG
-A1UEBhMCVUsxDzANBgNVBAcTBkxvbmRvbjESMBAGA1UEChMJVGVzdE9yZ0NBMREw
-DwYDVQQDEwh0ZXN0Lm9yZzCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIB
-AMlb3p6euGnp+SWOnIVv0JklDeB6QyuUitgdx8dqgwJO1aJZG5feLO4alrKVgOZ+
-BLhD9Z8zQmFztEP36Bu16Acu2dLho1MTF4Nyul9da0cLMNm+WnorwB/Ef68XluIf
-9gEJgBS8nNPyLcWR6COQ23OPdE/jo9agAzb7pX4X9atNEjHmdwD4KKEVJaKqyE0p
-ynNneDfJTS15GxVcGHoiqBcpwTSXmEzoCVein7YvJ9fCdXwteNO1fVvmKDdxtyIe
-8jgMYI7yVLRWfPfwMwf3NvCgB5sogYTwOIb2MZFFwTKK663s62U2xCtqNepI/Gdt
-rmW7ThvnbYKki8VOyK9dmuOgsVxeWrEwGbzb+2kPlp8oxk03VcFxDE8XTRK/EO6q
-F+VYnmEQkbUVgtq6fMcWYGZeqRdcMnLJxfWCTjhmni3bTSsgpxLbu0z54SNhFYIi
-/hAHyuatrGjRhiWmjmU5eDoem68Bdbxh6MgHFDdGzwgpQG+M2KimKXYYzicUdMm/
-WLx5hw6UTNvYyC8cMwZdhIYy7R96CnCqxQ6tW5I+rokTZy8F71bfdSE03rn2TFq6
-KpDRzBKYxa8BCjIdW/nvRHvEwIJo5X8uwUhVObcebPEjVeX9HPr135INHBxGddu0
-qzao3mdbWrQYHDXX5IhN1MqEQAuXEuXDU0deX+CjezDPAgMBAAGjLzAtMAwGA1Ud
-EwQFMAMBAf8wHQYDVR0OBBYEFEg8Tt7TRESjDuwjF6doAmhYu3H7MA0GCSqGSIb3
-DQEBDQUAA4ICAQBV2XmLkHIkSNT6KQRa1ehbH7I1FKOP199gQLqYUu0ojQIlPwWr
-c1huOpoajKF5Wc/5EfAkTIFUFOF569KoRTyVrlINcSRCNkhbYdIHT8b+MZO/LAFX
-ICnbbDnqJG/Yqjqx6dz4Sd431GJvpYtA6kdygTRuj0j6QD2WgClXXl0l+Ue4TzJw
-qvGFm8oyjCB7b/4FS/qCbDHs+vdME8VFYIzPmaUpgRcg2f1T2NwADc2gCQ1cyWjW
-anskt76eSct2zVgrNFLbqtVdxchJTC5gFZBAqadgnutSyvU1BatkzauW4Vo8/ZMk
-z1QlMgTwfCdL+msPm/4XpLPpUbSe9GLED6BT9MCG5RsTDSwkEotaqjB+vcMAuivX
-6cmnNCY5ReX36iP31xgd5oiwqnPfhiGUFKJBKv8PiCl/ALzVIdaO/lO5thUvAlr8
-p6K/x2g918IDBb4LP9EjMFNwyJFBMSDkUrWk+o9VO761qYcshswkbg1TyAlsFZdc
-4ILxM99nhIvD+D7q6iJjHPz9GxLmLpXvNE30JbFLsrUuccjBaUdE6b5xm8be83cA
-YrvRicgZmPcr1hm5mKFYhUEOaa2zLWVvBMEQ4hnNLfSrH+2AVQFB/ryv7KDoeqU3
-R888g2ibQJvSXeUdXXq9uUloa5/qa0Lqysk8a2z369vM1VR0DbZ6Aus6vg==
------END CERTIFICATE-----
