--- extension-impl/registry-consul/src/test/resources/ssl/key.key
@@ -1,56 +1,0 @@
-Bag Attributes
-    friendlyName: key
-    localKeyID: 54 69 6D 65 20 31 35 33 32 34 39 38 39 36 36 33 32 39 
-Key Attributes: <No Attributes>
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQC2mhwSB5PzXikf
-wL7uAO6nml8GTEGE7LEq97H+P/UCXj8PqXBOEmUno3EZQqcqC3mNuvTg7zY1pMCt
-ihPM53Zi/4U9oXudv8jt5FnOJmRs1bYLckFGS0DHOlXGZUquofWffM9nBLsxzgsd
-IKNTCSvyDgZoyVbZeLmYJqm/k8K/Yme7wP3GDBeuvgLgV0yfz751onhSOT2p2GBZ
-49cdvb8a9A2YRguKikBGlaad9UZM5qSRcvlmZZ2X2cRK32WjBjc1qSiH6e17NzoT
-sUgup/wWuDt2FuRfmS6P8/SkK1uiGSTU/gcfFl4XCVvHPYIJBGMw/G2NsiP/cBfq
-9k8xTt+wcuXemqT1GQ/FsnjV1fm/FXqaZXgCNJLm8ZFEXBaYjyoHUaILCs5lqSnS
-IdklZbv7h4Q0JVpzoapDzkWVWJ+S+GEVwqq8iwPlSaHfofuAhAqKK6SblQRBWqXN
-UP4N4WIakLKg7ETCrIkCxJL93exucRwS4a1+dsQnK2kE8OA5Mk9Ly6U6gBIUGg8G
-xJCakv0UhlXKwqECUJ5D0oIva/WAB9vMZ/x+Da9yWJrctoyjvWueaRzjqQoG3OB1
-X43OnhdO00pwnK3mNtVjp38vRxfVsAjsxlHmv0ORcSnuM2Ukic74o/fgPOAKlz62
-zXauM7FZKkGBB7UJU2GdOKV4Kp47iwIDAQABAoICACQOiE5zGp7qz/oKKGl826NS
-00K+Gg3yYwWWBZqr7Gh8Tb18+sUf0sOh/Xo8dB/URjtbWHF2KuEq2DUsV76VUgY4
-aIoD8hHzRezbs3QGlrv680iNViBYScFNXpSfecDEEGlezuGLyJ8MDZvBzx5daEUS
-yec1IqYstZE6yTggKEt8Vfl4IWRTDPYfZRgBhQrljblSzKmHzgpUxKnbWKXxyuO6
-n1ibNDo2FH0bndwBC9GxlXt3FZJHrHfIs38JX9tyJcK91VgJNDCpdhlfX15/I0Ml
-O/CUkqHh8JTSWlN0kFKQ3noEc0mLNCceXPYZyiyXpjKNFeDt9TfCXaay3m5hIcAy
-nBqovP4UxSE7uIpzoGtspJdM9WrTgpEEMOf96WFoBwCo3w3RAlgDcKDNnm8hdC+j
-BvY1kjE9jFi6nRHLsmyJnfAj70sxpHLo9ZESbHqzcZ1BfEF8fqm++NSFvoqbxbJQ
-f4+mTsZTbnfmgz/Y+Iwxh93XTT5ALloLbrQqNBi+DYYQl63rJg6F2VNF1dyvrFhP
-DBjApHO+kkivqA6HujMlees6zia7eKrJ0zHPa+H1obiwO2qcFAx6ERUR762DuCQ+
-PJkeZmDTngNBCa+JcxOTFwGxJC89j8eky6mC3B8GUcrWiyK1PM6PNLIMUOsiqfOI
-zwe1uiQJfiG1pY2LQwiZAoIBAQD5hjsCG6It6YMVyMcIBhesubrXA0OPRrVg65rf
-ZzD1qtLdY2ylgUqK+r4+4xdo/argEr83Yw8srsQzF7Wcyff94MNWgmts7y234PDz
-qOADAbbxJ9SHIkULkVM/DrruIZJUmUqcnHw6eOtwGAkKxN11PwmXo3CYZ3Yz1Er0
-NN6BW+dVB9uje42OJqvDIQwWVaUuU7rhuG96AHo0hsi+AKSSd8zYh/Tbvo17QGj+
-EzhL6qD+cs+KtERSEDrJjKXFnVaj2MKLLrr/cf5yeUXF1CnGQ/i/1PZ4px+KOeA8
-LyCVElH7JrkTZOnwSwucZNgImll+M5uVBysDXqM6gF3OzUSvAoIBAQC7V0QSPiLA
-/VnMtAzm8GienqbAT/5ssXy94+mk2vBTSDyWRoSZPuv2MQ0TntIRzhlwI/IjvvU1
-J16hVd2Awr+YJ7P4cYcftcqrt9T79OCPWbhSpqgjx1CHSg6F9J5fg274HEMQ1z6/
-Cf3e7qKG6BUWSoXVe0aigmQ3hYq1qpBliUJbUgzPGNtBWhmfZKBmHyL/N9BdCXu7
-fpscUMbiUzNWRaiLX2DTykYz3FkzTq7PnRL65LxQ2G3ulUuZIpUV2OeKhiCqgFns
-BY05rE7ESzAxVFaOl1wh/nsb8pmyLcTDZ205s5fzvh0cZJ6SCskFgI1RC6aQxUuJ
-ZWh/+DawtaXlAoIBAQD3fMWGP7ZJ53QW5T9TnAdbUBQbMqeNskowST8hGg6be8UN
-xJF8TpKK6kxLv532irMy8CUsYFoqrWT4cCm3yzLnYkkcwtqd/1TPLpKOeg9FrCux
-YRoQtlLZOFjjEE8j5bRnqGsyjptAu7J5Sb2eKsikg9OVzYvd/bBlk8iMEuwmTetS
-kSquKdurBHwUW/DLLkNZgb64JuCal2KeOplcqvKQO7a3hZG9anCVsLryVkneMmdg
-OKvVHRsucd59Db1WFcpbVKEQv1bGCmCfVd14bAjLF5vfWPDllo1X4jOgPfYJJVSe
-0Dd9HvhRV326ruL/FWhyHenQlH17KCCp46wnc70NAoIBAEbNr4N5imcIGaVdMzch
-C1lPutQaJ51q4ytddROCT6waNBPTqQ/GESYl2bxdnSVvkHuqijxqUoB5VbGCI0oj
-0IewEbvHJHuemG+X9d4X6bsj/J74L7n9/Xh0S9Li5exRX2JURCCcSTGolBNqe3Cj
-TDz1g2lwwvxubXPye783LIXDf5nGA+NNkKg9Zkuwm1KdV/u/W3eeOu5c5kCoYtJt
-AY0MnKu4K39ABpJZKDlU9oqu4dqDEo03jyzJrh29TT1JkhDvqm7mPZYPbxWqAsb7
-Z8ZtCp2vpPmQFXU1llMaKzBy6XwSXU6YvqXZfXjSrLDwqfhEf/pdJrRR6FHdPMZC
-kOUCggEBAI9LBN9rXfTCOjFCmQ7NEBLUsb0fHwS3J1JxI21aY4Efys5zqGNt+V5M
-rrCQAdtjutVvJkdzF/ZRbPtYEmQ0aJDjQEB+TK0XTyvZx1ycEnP6RRY2WR5KKMPO
-ZTSaNIeYXvyW6kL+2u4cnn/jTGu1PnnQ2zbyFCpMTzQdquJBXlY5XyCs1Z15NOYe
-vbsIRCwLg5Dhb2tFjgpDMl5VohA4/uC0SRWWn0m9UB51lGcl6OjUwTTfBoiYTkh+
-2m5tjLVwYdbVpZVjHXsHQinaWqECmzt7li/pZr7vplJl0UZ4PVgDcCQK1Z2mvlY9
-B11TuC9bkEJXvyeg2H9OKsYNyzA3Ubk=
------END PRIVATE KEY-----
