+++ extension-impl/registry-zk/src/main/java/com/alipay/sofa/rpc/registry/zk/ZookeeperRegistryHelper.java
@@ -35,6 +35,7 @@ import com.alipay.sofa.rpc.config.ConsumerConfig;
 import com.alipay.sofa.rpc.config.ProviderConfig;
 import com.alipay.sofa.rpc.config.ServerConfig;
 import com.alipay.sofa.rpc.context.RpcRuntimeContext;
+import com.alipay.sofa.rpc.utils.RegistryUtils;
 import org.apache.curator.framework.recipes.cache.ChildData;
 
 import java.io.UnsupportedEncodingException;
@@ -52,102 +53,7 @@ import java.util.concurrent.ConcurrentHashMap;
  *
  * @author <a href=mailto:zhanggeng.zg@antfin.com>GengZhang</a>
  */
-public class ZookeeperRegistryHelper {
-
-    /**
-     * Convert provider to url.
-     *
-     * @param providerConfig the ProviderConfig
-     * @return the url list
-     */
-    static List<String> convertProviderToUrls(ProviderConfig providerConfig) {
-        List<ServerConfig> servers = providerConfig.getServer();
-        if (servers != null && !servers.isEmpty()) {
-            List<String> urls = new ArrayList<String>();
-            for (ServerConfig server : servers) {
-                StringBuilder sb = new StringBuilder(200);
-                String host = server.getVirtualHost(); // 虚拟ip
-                if (host == null) {
-                    host = server.getHost();
-                    if (NetUtils.isLocalHost(host) || NetUtils.isAnyHost(host)) {
-                        host = SystemInfo.getLocalHost();
-                    }
-                }
-                sb.append(server.getProtocol()).append("://").append(host).append(":")
-                    .append(server.getPort()).append(server.getContextPath()).append("?version=1.0")
-                    .append(
-                        getKeyPairs(RpcConstants.CONFIG_KEY_UNIQUEID, providerConfig.getUniqueId()))
-                    .append(
-                        getKeyPairs(RpcConstants.CONFIG_KEY_TIMEOUT, providerConfig.getTimeout()))
-                    .append(getKeyPairs("delay", providerConfig.getDelay()))
-                    .append(getKeyPairs("id", providerConfig.getId()))
-                    .append(
-                        getKeyPairs(RpcConstants.CONFIG_KEY_DYNAMIC, providerConfig.isDynamic()))
-                    .append(getKeyPairs(ProviderInfoAttrs.ATTR_WEIGHT, providerConfig.getWeight()))
-                    .append(getKeyPairs("accepts", server.getAccepts()))
-                    .append(getKeyPairs(ProviderInfoAttrs.ATTR_START_TIME, RpcRuntimeContext.now()))
-                    .append(
-                        getKeyPairs(RpcConstants.CONFIG_KEY_APP_NAME, providerConfig.getAppName()))
-                    .append(getKeyPairs(RpcConstants.CONFIG_KEY_SERIALIZATION,
-                        providerConfig.getSerialization()))
-                    .append(convertMap2Pair(providerConfig.getParameters()));
-                addCommonAttrs(sb);
-                urls.add(sb.toString());
-            }
-            return urls;
-        }
-        return null;
-    }
-
-    /**
-     * Convert consumer to url.
-     *
-     * @param consumerConfig the ConsumerConfig
-     * @return the url list
-     */
-    static String convertConsumerToUrl(ConsumerConfig consumerConfig) {
-        StringBuilder sb = new StringBuilder(200);
-        String host = SystemInfo.getLocalHost();
-        sb.append(consumerConfig.getProtocol()).append("://").append(host).append("?version=1.0")
-            .append(getKeyPairs(RpcConstants.CONFIG_KEY_UNIQUEID, consumerConfig.getUniqueId()))
-            .append(getKeyPairs("pid", RpcRuntimeContext.PID))
-            .append(getKeyPairs(RpcConstants.CONFIG_KEY_TIMEOUT, consumerConfig.getTimeout()))
-            .append(getKeyPairs("id", consumerConfig.getId()))
-            .append(getKeyPairs(RpcConstants.CONFIG_KEY_GENERIC, consumerConfig.isGeneric()))
-            .append(getKeyPairs(RpcConstants.CONFIG_KEY_APP_NAME, consumerConfig.getAppName()))
-            .append(getKeyPairs(RpcConstants.CONFIG_KEY_SERIALIZATION,
-                consumerConfig.getSerialization()))
-            .append(getKeyPairs(ProviderInfoAttrs.ATTR_START_TIME, RpcRuntimeContext.now()))
-            .append(convertMap2Pair(consumerConfig.getParameters()));
-        addCommonAttrs(sb);
-        return sb.toString();
-    }
-
-    /**
-     * Gets key pairs.
-     *
-     * @param key   the key
-     * @param value the value
-     * @return the key pairs
-     */
-    private static String getKeyPairs(String key, Object value) {
-        if (value != null) {
-            return "&" + key + "=" + value.toString();
-        } else {
-            return "";
-        }
-    }
-
-    /**
-     * 加入一些公共的额外属性
-     *
-     * @param sb 属性
-     */
-    private static void addCommonAttrs(StringBuilder sb) {
-        sb.append(getKeyPairs("pid", RpcRuntimeContext.PID));
-        sb.append(getKeyPairs("language", "java"));
-        sb.append(getKeyPairs(RpcConstants.CONFIG_KEY_RPC_VERSION, Version.RPC_VERSION + ""));
-    }
+public class ZookeeperRegistryHelper extends RegistryUtils {
 
     /**
      * 转换 map to url pair
@@ -312,52 +218,6 @@ public class ZookeeperRegistryHelper {
             }
         }
         return attribute;
-    }
-
-    /**
-     * Read the warmup weight parameter,
-     * decide whether to switch the state to the preheating period,
-     * and set the corresponding parameters during the preheating period.
-     *
-     * @param providerInfo
-     */
-    static void processWarmUpWeight(ProviderInfo providerInfo) {
-
-        String warmupTimeStr = providerInfo.getStaticAttr(ProviderInfoAttrs.ATTR_WARMUP_TIME);
-        String warmupWeightStr = providerInfo.getStaticAttr(ProviderInfoAttrs.ATTR_WARMUP_WEIGHT);
-        String startTimeStr = providerInfo.getStaticAttr(ProviderInfoAttrs.ATTR_START_TIME);
-
-        if (StringUtils.isNotBlank(warmupTimeStr) && StringUtils.isNotBlank(warmupWeightStr)
-            && StringUtils.isNotBlank(startTimeStr)) {
-
-            long warmupTime = CommonUtils.parseLong(warmupTimeStr, 0);
-            int warmupWeight = CommonUtils.parseInt(warmupWeightStr,
-                Integer.parseInt(providerInfo.getStaticAttr(ProviderInfoAttrs.ATTR_WEIGHT)));
-            long startTime = CommonUtils.parseLong(startTimeStr, 0);
-            long warmupEndTime = startTime + warmupTime;
-
-            // set for dynamic
-            providerInfo.setDynamicAttr(ProviderInfoAttrs.ATTR_WARMUP_WEIGHT, warmupWeight);
-            providerInfo.setDynamicAttr(ProviderInfoAttrs.ATTR_WARM_UP_END_TIME, warmupEndTime);
-            providerInfo.setStatus(ProviderStatus.WARMING_UP);
-        }
-
-        // remove from static
-        providerInfo.getStaticAttrs().remove(ProviderInfoAttrs.ATTR_WARMUP_TIME);
-        providerInfo.getStaticAttrs().remove(ProviderInfoAttrs.ATTR_WARMUP_WEIGHT);
-
-    }
-
-    static String buildProviderPath(String rootPath, AbstractInterfaceConfig config) {
-        return rootPath + "sofa-rpc/" + config.getInterfaceId() + "/providers";
-    }
-
-    static String buildConsumerPath(String rootPath, AbstractInterfaceConfig config) {
-        return rootPath + "sofa-rpc/" + config.getInterfaceId() + "/consumers";
-    }
-
-    static String buildConfigPath(String rootPath, AbstractInterfaceConfig config) {
-        return rootPath + "sofa-rpc/" + config.getInterfaceId() + "/configs";
     }
 
     static String buildOverridePath(String rootPath, AbstractInterfaceConfig config) {
